pyproject.toml
[project]
name = "resolv"
version = "0.1.0"
description = ""
authors = [
    {name = "Romanos Chiliarchopoulos",email = "romanoshiliarhopoulos@gmail.com"}
]
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "requests (>=2.33.1,<3.0.0)",
    "bpy (==4.2.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.resolv]
blender_path = ".blender/blender"
[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)"
]
